Remove duplicate mock_supervisor_comms pytest fixture definitions
